Upload Client Separation so no conflicting of open/close values for separate forms and working buttons now
