+++ website/src/Pages/TeamBandit/OrganizerPages/TeamBanditPages/Clients/Components/FormDialogAddClient.js
@@ -22,6 +22,8 @@ import TableRow from "@mui/material/TableRow";
 import AddIcon from "@mui/icons-material/Add";
 import FileUploadIcon from "@mui/icons-material/FileUpload";
 
+import FormDialogUploadClients from "./FormDialogUploadClients";
+
 // Need this to change color schemes
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 
@@ -313,267 +315,16 @@ const FormDialogAddClient = ({ setClientsChange }) => {
                     {" "}
                     Upload Client List{" "}
                 </Button>
+                
                 <Dialog
-                    maxHeight={"lg"}
-                    maxWidth={"lg"}
-                    open={open}
-                    onClose={handleToggle}
+                maxHeight={"lg"}
+                maxWidth={"lg"}
+                open={open}
+                onClose={handleToggle}
                 >
-                    <div className={styles.csv}>
-                        <div className={styles.appHeader}>
-                            <div className={styles.uploader}>
-                                <div
-                                    className={`${styles.appLink} ${
-                                        highlighted
-                                            ? styles.appHighlighted
-                                            : styles.appNothing
-                                    }`}
-                                    onDragEnter={() => {
-                                        setHighlighted(true);
-                                    }}
-                                    onDragLeave={() => {
-                                        setHighlighted(false);
-                                    }}
-                                    onDragOver={(e) => {
-                                        e.preventDefault();
-                                    }}
-                                    onDrop={(e) => {
-                                        e.preventDefault();
-                                        setHighlighted(false);
-
-                                        Array.from(e.dataTransfer.files)
-                                            .filter(
-                                                (file) =>
-                                                    file.type ===
-                                                    "application/vnd.ms-excel"
-                                            )
-                                            .forEach(async (file) => {
-                                                const text = await file.text();
-                                                const result = parse(text, {
-                                                    header: true,
-                                                });
-                                                setContacts((existing) => [
-                                                    ...existing,
-                                                    ...result.data,
-                                                ]);
-                                            });
-                                    }}
-                                >
-                                    DRAG AND DROP FILE HERE
-                                </div>
-
-                                <p className={styles.p}> OR </p>
-
-                                <label htmlFor="contained-button-file">
-                                    <ThemeProvider theme={theme}>
-                                        <Input
-                                            accept="text/csv"
-                                            id="contained-button-file"
-                                            multiple
-                                            type="file"
-                                            onChange={(event) => {
-                                                event.preventDefault();
-                                                const loadCSV = Promise.resolve(
-                                                    event.target.files[0].text()
-                                                );
-
-                                                loadCSV.then(function (
-                                                    resultOfPromise
-                                                ) {
-                                                    const parsedResult = parse(
-                                                        resultOfPromise,
-                                                        { header: true }
-                                                    );
-                                                    setContacts((existing) => [
-                                                        ...existing,
-                                                        ...parsedResult.data,
-                                                    ]);
-                                                });
-                                            }}
-                                        />
-
-                                        <Button
-                                            color="secondary"
-                                            variant="contained"
-                                            component="span"
-                                        >
-                                            Upload .CSV File
-                                        </Button>
-                                    </ThemeProvider>
-                                </label>
-                            </div>
-                            <div className={styles.tablepad}>
-                                <TableContainer component={Paper}>
-                                    <Table
-                                        sx={{ minWidth: 1000 }}
-                                        aria-label="simple table"
-                                    >
-                                        <TableHead>
-                                            <TableRow>
-                                                <TableCell
-                                                    style={{
-                                                        backgroundColor: "#F2C12E",
-                                                        color: "black",
-                                                    }}
-                                                >
-                                                    First Name
-                                                </TableCell>
-                                                <TableCell
-                                                    style={{
-                                                        backgroundColor: "#F2C12E",
-                                                        color: "black",
-                                                    }}
-                                                >
-                                                    Last Name
-                                                </TableCell>
-                                                <TableCell
-                                                    style={{
-                                                        backgroundColor: "#F2C12E",
-                                                        color: "black",
-                                                    }}
-                                                    align="right"
-                                                >
-                                                    Email
-                                                </TableCell>
-                                                <TableCell
-                                                    style={{
-                                                        backgroundColor: "#F2C12E",
-                                                        color: "black",
-                                                    }}
-                                                    align="right"
-                                                >
-                                                    Organization
-                                                </TableCell>
-                                                <TableCell
-                                                    style={{
-                                                        backgroundColor: "#F2C12E",
-                                                        color: "black",
-                                                    }}
-                                                    align="right"
-                                                >
-                                                    Phone Number
-                                                </TableCell>
-                                                <TableCell
-                                                    style={{
-                                                        backgroundColor: "#F2C12E",
-                                                        color: "black",
-                                                    }}
-                                                    align="right"
-                                                >
-                                                    Notes
-                                                </TableCell>
-                                            </TableRow>
-                                        </TableHead>
-                                        <TableBody>
-                                            {contacts.map((contact) => (
-                                                <TableRow
-                                                    key={contact.email}
-                                                    sx={{
-                                                        "&:last-child td, &:last-child th":
-                                                            { border: 0 },
-                                                    }}
-                                                >
-                                                    <TableCell
-                                                        style={{
-                                                            backgroundColor:
-                                                                "#003466",
-                                                            color: "white",
-                                                        }}
-                                                        component="th"
-                                                        scope="row"
-                                                    >
-                                                        {contact.firstName}
-                                                    </TableCell>
-                                                    <TableCell
-                                                        style={{
-                                                            backgroundColor:
-                                                                "#003466",
-                                                            color: "white",
-                                                        }}
-                                                        component="th"
-                                                        scope="row"
-                                                    >
-                                                        {contact.lastName}
-                                                    </TableCell>
-                                                    <TableCell
-                                                        style={{
-                                                            backgroundColor:
-                                                                "#003466",
-                                                            color: "white",
-                                                        }}
-                                                        align="right"
-                                                    >
-                                                        {contact.email}
-                                                    </TableCell>
-                                                    <TableCell
-                                                        style={{
-                                                            backgroundColor:
-                                                                "#003466",
-                                                            color: "white",
-                                                        }}
-                                                        align="right"
-                                                    >
-                                                        {contact.organization}
-                                                    </TableCell>
-                                                    <TableCell
-                                                        style={{
-                                                            backgroundColor:
-                                                                "#003466",
-                                                            color: "white",
-                                                        }}
-                                                        align="right"
-                                                    >
-                                                        {contact.phone}
-                                                    </TableCell>
-                                                    <TableCell
-                                                        style={{
-                                                            backgroundColor:
-                                                                "#003466",
-                                                            color: "white",
-                                                        }}
-                                                        align="right"
-                                                    >
-                                                        {contact.notes}
-                                                    </TableCell>
-                                                </TableRow>
-                                            ))}
-                                        </TableBody>
-                                    </Table>
-
-                                </TableContainer>
-                                
-                                <Button
-                                    variant="contained"
-                                    style={{
-                                        backgroundColor: "#003466",
-                                    }}
-
-                                    onClick={handleToggle}
-                                >
-                                {" "}
-                                Cancel
-                                {" "}
-                                </Button>
-
-                                {" "}
-                                
-                                <Button
-                                    variant="contained"
-                                    color="secondary"
-                                    onClick={addList}
-                                >
-                                    {" "}
-                                    Commit Changes
-                                    {" "}
-                                </Button>
-                                
-                                {" "}
-                                
-
-                            </div>
-                        </div>
-                    </div>
+                    <FormDialogUploadClients setClientsChange={setClientsChange} setOpen={setOpen} open={open}/>
                 </Dialog>
+                
             </div>
            );
 };
